pyproject.toml
[tool.poetry]
name = "magic_paper"
version = "0.1.0"
description = ""
authors = ["DM1122"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Pillow = "^9.2.0"
numpy = "^1.23.2"
inky = {extras = ["rpi"], version = "^1.3.2"}

[tool.poetry.dev-dependencies]
ipykernel = "^6.13.0"
nbformat = "^5.3.0"
pre-commit = "^2.16.0"
pylint = "^v2.13.5"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = false
ignore_missing_imports = true

[tool.isort]
profile = "black"
sections="FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
import_heading_stdlib="stdlib"
import_heading_thirdparty="external"
import_heading_firstparty="project"
known_first_party = ["magic_paper"]

[tool.black]
line-length = 88

[tool.docformatter] # https://github.com/myint/docformatter/pull/77
blank = true
force-wrap = false
in-place = true
make-summary-multi-line = false
pre-summary-newline = false
recursive = true
wrap-descriptions = 88
wrap-summaries = 80

[tool.pylint.master]
ignore-patterns="^((?!\\.py).)*$"  # ignore files that are not python files
recursive="yes"
disable = ["invalid-name","too-many-arguments", "too-many-locals", "arguments-differ", "unnecessary-lambda", "too-many-instance-attributes", "protected-access", "duplicate-code", "too-few-public-methods", "logging-fstring-interpolation", "not-callable", "too-many-statements", "pointless-statement", "not-context-manager","fixme","too-many-branches","too-many-lines","line-too-long","anomalous-backslash-in-string","too-many-function-args"]
max-line-length = 88
generated-members=["numpy.*","torch.*","pygame.*"] # https://github.com/pytorch/pytorch/issues/701

[tool.pytest.ini_options]
addopts = ["--cov", "--cov-report=xml:logs/coverage.xml", "--durations=0", "--new-first", "--strict-config", "--failed-first", "--verbose"] 
console_output_style = "count"
faulthandler_timeout = 10
log_auto_indent = true
log_cli = false
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_level = "INFO"
log_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "logs/pytest.log"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(funcName)s:%(lineno)s)"
log_file_level = "DEBUG"
log_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(funcName)s:%(lineno)s)"
log_level = "DEBUG"
testpaths = ["tests"]

[tool.coverage.run]
data_file = "logs/.coverage"
omit = ["tests/*"]